% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_mics_sib_histories.R
\name{build_mics_ego_cols}
\alias{build_mics_ego_cols}
\title{construct the respondent-level columns MICS does not supply}
\usage{
build_mics_ego_cols(
  df,
  survey,
  id.vars = c("hh1", "hh2", "ln"),
  doi.var = "wdoi",
  doi.ym = c("wm6y", "wm6m"),
  dob.var = "wdob",
  verbose = TRUE
)
}
\arguments{
\item{df}{the (renamed) MICS sibling file}

\item{survey}{the survey id}

\item{id.vars}{columns identifying a respondent}

\item{doi.var}{CMC date-of-interview column, used when present}

\item{doi.ym}{year and month columns, used when \code{doi.var} is absent}

\item{dob.var}{CMC respondent date-of-birth column}

\item{verbose}{report what was constructed?}
}
\value{
\code{df} with \code{caseid}, \code{survey}, \code{doi}, \code{age}, \code{psu} and \code{sex} added
}
\description{
construct the respondent-level columns MICS does not supply
}
