% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sibship_visibility.R
\name{calculate_sib_ind_visibility}
\alias{calculate_sib_ind_visibility}
\title{given a sib dataset, calculate individual visibility weight for each sib}
\usage{
calculate_sib_ind_visibility(
  df,
  sib.frame.indicator = ".sib.in.F",
  num.sibs.on.frame.var = "y.F",
  varname = "ind_vis"
)
}
\arguments{
\item{df}{Dataset on sibling reports (possibly by cell)}

\item{sib.frame.indicator}{String with the name of the column in \code{df} containing a 0/1 coded variable indicating whether or not each sib is in the frame population}

\item{num.sibs.on.frame.var}{String with the name of the column in \code{df} containing the number of sibs on the frame for each ego}

\item{varname}{String with the name of the new column to create with the individual visibility of each sib}
}
\value{
\code{df} with a new column that has the individual visibility of each sibling
}
\description{
given a sib dataset, calculate individual visibility weight for each sib
}
