% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling_estimator.R
\name{get_agg_est_from_ec}
\alias{get_agg_est_from_ec}
\title{helper function for calculating aggregate visibility estimate from ego X cell data}
\usage{
get_agg_est_from_ec(ec_dat, wgt_var, cell_vars)
}
\arguments{
\item{ec_dat}{the ego X cell data}

\item{wgt_var}{string with the name of the column that has sampling weights}

\item{cell_vars}{vector of strings with the names of variables to group by (the cells)}
}
\value{
a tibble with the individual visibility ASDR estimates (not including the respondents' exposures)
}
\description{
helper function for calculating aggregate visibility estimate from ego X cell data
}
