% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esc_reports.R
\name{get_esc_reports}
\alias{get_esc_reports}
\title{Get ego X sibling X cell reports}
\usage{
get_esc_reports(sib.dat, sib.id, ego.id, cell.config)
}
\arguments{
\item{sib.dat}{The long-form sibling history dataset (likely produced by TODO)}

\item{sib.id}{String with the name of the column of \code{sib.dat} that has the sibling ID}

\item{ego.id}{String with the name of the column of \code{sib.dat} that has the ID of the survey respondent}

\item{cell.config}{An object containing the configuration of cells; see TODO for more information}
}
\value{
A tbl with a row for each respondent-sibling X cell
}
\description{
Produces a long-form dataset that has a row for each respondent-sibling X cell combination
}
\examples{
  # TODO
}
