% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sibship_visibility.R
\name{get_sibship_info}
\alias{get_sibship_info}
\title{calculate number of sibs on frame for each respondent}
\usage{
get_sibship_info(sib.dat, ego.id, sib.frame.indicator)
}
\arguments{
\item{sib.dat}{The long-form sibling dataset (likely produced by a prep function such as \link{prep_dhs_sib_histories})}

\item{ego.id}{String with the name of the column in \code{sib.dat} containing the survey respondent ID}

\item{sib.frame.indicator}{String with the name of the column in \code{sib.dat} containing a 0/1 coded variable indicating whether or not each sib is in the frame population}
}
\value{
A tibble with a row for each survey respondent (each unique value of \code{ego.id}), and the number of sibs the respondent reported on the frame, including and not including herself
}
\description{
this quantity, y.F, is related to the visibility of each respondent
}
\examples{
  # TODO write example code
}
