% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_config.R
\name{make.even.age.groups}
\alias{make.even.age.groups}
\title{make an age.groups object with evenly-sized intervals}
\usage{
make.even.age.groups(min.age, max.age, width)
}
\arguments{
\item{min.age}{the first age}

\item{max.age}{the highest age}

\item{width}{the width of each group}
}
\value{
see make.age.groups
}
\description{
see make.age.groups for more
}
