% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{occ.exp}
\alias{cell_config}
\alias{make.age.groups}
\alias{make.even.age.groups}
\alias{make.time.periods}
\alias{nmx_to_nqx}
\alias{q15_to_50}
\alias{get_visibility}
\alias{sib_ic_checks}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{networkreporting}{\code{\link[networkreporting]{cell_config}}, \code{\link[networkreporting]{get_visibility}}, \code{\link[networkreporting]{make.age.groups}}, \code{\link[networkreporting]{make.even.age.groups}}, \code{\link[networkreporting]{make.time.periods}}, \code{\link[networkreporting]{nmx_to_nqx}}, \code{\link[networkreporting]{occ.exp}}, \code{\link[networkreporting]{q15_to_50}}, \code{\link[networkreporting]{sib_ic_checks}}}
}}

