% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ic_reports.R
\name{sib_ic_checks}
\alias{sib_ic_checks}
\title{get calculate internal consistency checks for sibling reports}
\usage{
sib_ic_checks(esc.dat, ego.dat, ego.id, sib.id, sib.frame.indicator,
  sib.cell.vars, ego.cell.vars, boot.weights)
}
\arguments{
\item{esc.dat}{The ego X sibling X cell dataset (see \link{\code{get_esc_reports}})}

\item{ego.dat}{The ego dataset, containing one row for each survey respondent}

\item{ego.id}{String with the name of the column in \code{esc.dat} containing the survey respondent ID}

\item{sib.id}{String with the name of the column in \code{esc.dat} containing the unique sibling ID}

\item{sib.frame.indicator}{String with the name of the column in \code{sib.dat} containing a 0/1 coded variable indicating whether or not each sib is in the frame population}

\item{sib.cell.vars}{see Details}

\item{ego.cell.vars}{see Details}

\item{boot.weights}{Dataframe with bootstrap resampled weights. See Details}
}
\value{
A tibble with a row for each reported sibling in the frame population, the sibling's cell info, and the cell info of the survey respondent who reported each sibling
}
\description{
get calculate internal consistency checks for sibling reports
}
\section{Details}{

The \code{sib.cell.vars} and \code{ego.cell.vars} arguments should have a vector with Strings containing the names of
columns in \code{esc.dat} identifying the cells to group reports by (typically age and sex, and possibly other variables).
Note that, unlike the \code{cell.vars} argument in \link{\code{get_ec_reports}}, for this function \code{sib.cell.vars} and
\code{ego.cell.vars} MUST have the age variable listed FIRST.
The goal of \code{get_ic_reports} is to find the cell that each sibling was in at the time of the interview, meaning the oldest age group to which the
sibling contributed any exposure.
\code{boot.weights} is assumed to have a column that is named whatever the \code{ego.id} is,
and then a series of columns named \code{boot_weight_1}, ..., \code{boot_weight_M}.
}

\examples{
  # TODO write example code
}
