% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sibhist_varmap_dhs4}
\alias{sibhist_varmap_dhs4}
\title{DHS sibling history variable map for DHS phase 4}
\format{
A data frame with columns for original and standardized variable names.
}
\description{
A data frame mapping variable names used in DHS phase 4 sibling history
modules to the standardized names used by this package.
}
