% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sibhist_varmap_mics5}
\alias{sibhist_varmap_mics5}
\title{MICS sibling history variable map for MICS5}
\format{
A data frame with columns \code{orig.varname}, \code{new.varname}, \code{sibvar},
\code{description} and \code{comments}.
}
\description{
Identical to \link{sibhist_varmap_mics4}; MICS4 and MICS5 use the same roster
numbering.
}
