% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sibhist_varmap_mics7}
\alias{sibhist_varmap_mics7}
\title{MICS sibling history variable map for MICS7}
\format{
A data frame with columns \code{orig.varname}, \code{new.varname}, \code{sibvar},
\code{description} and \code{comments}.
}
\description{
Identical to \link{sibhist_varmap_mics6}; the MICS7 maternal mortality module is
unchanged from MICS6.
}
