% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dhs_sib_histories.R
\name{add_maternal_deaths}
\alias{add_maternal_deaths}
\title{take a prepared DHS sibling dataset and add maternal death info}
\usage{
add_maternal_deaths(sib_df, keep_missing = FALSE, verbose = TRUE)
}
\arguments{
\item{sib_df}{the prepped DHS dataset (probably from \link{prep_dhs_sib_histories})}

\item{keep_missing}{should we keep reported sibs that are missing sex or survival status?}

\item{verbose}{report detailed summaries?}
}
\value{
a dataframe with columns \code{sib.preg_related.death.date} and \code{sib.maternal.death.date} added
}
\description{
take a prepared DHS sibling dataset and add maternal death info
}
\section{Details}{


This function checks to see if there is a column called \code{sib.died.accident}.
If so, then it is possible to estimate whether or not each death is maternal;
the resulting death dates are in the column \code{sib.maternal.death.date}.

For siblings, you should be sure to include
\itemize{
\item \code{sib.death.date} (the date of the sibling's death)
\item \code{sib.alive} (whether or not the sib is alive)
\item \code{sib.sex} (the sex of the sibling).
}

Returns a dataframe TODO
}

\examples{
  # TODO - write example code
}
