% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_config.R
\name{agenames}
\alias{agenames}
\title{make labels for age groups}
\usage{
agenames(aw, min.age, max.age)
}
\arguments{
\item{aw}{the width of each group}

\item{min.age}{the first age}

\item{max.age}{the highest age}
}
\value{
see make.age.groups
}
\description{
helper function; see make.even.age.groups for more
}
