% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternal_estimators.R
\name{aggregate_maternal_estimates}
\alias{aggregate_maternal_estimates}
\title{calculate total rate based on point estimates}
\usage{
aggregate_maternal_estimates(estimates, ego.dat, sib.dat, only_females = TRUE)
}
\arguments{
\item{estimates}{the output of \link{sibling_estimator}}

\item{ego.dat}{the prepped ego data}

\item{sib.dat}{the prepped sibling data}

\item{only_females}{only keep female estimates? Defaults to TRUE; see Details, below}
}
\value{
either a tibble with the estimates aggregated across age groups
OR, if there are bootstrap results, then a list with three entries:
\itemize{
\item \code{point} - the point estimates (should be the same as running w/out bootstraps)
\item \code{boot_summ} - estimates w/ confidence intervals calculated from bootstrap reps
\item \code{boot} - results for each bootstrap rep
}
}
\description{
calculate total rate based on point estimates
}
\section{Details}{


This function aggregates age-specific maternal mortality quantities
like the maternal death rate or the pregnancy-related death rate.
It uses the age distribution of the survey respondents as the reference
population.

Note that, by default, this will only consider estimates for females aged 15-49,
as this is what makes sense for aggregating maternal estimates.
It might occasionally be useful to run this function for all-cause mortality;
in that case, setting the parameter \code{only_female=FALSE} will include
male estimates, too (but will still restrict to ages 15-49)
}

