% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sibship_visibility.R
\name{get_ego_age_distn}
\alias{get_ego_age_distn}
\title{get_ego_age_distribution}
\usage{
get_ego_age_distn(ego.dat, only_females = TRUE)
}
\arguments{
\item{ego.dat}{the ego dataset (probably from \link{prep_dhs_sib_histories})}

\item{only_females}{should only females be used to calculate age distribution? (default: True)}
}
\value{
dataframe with distribution of respondent ages by 5-year category,
typically used in calculating the maternal or pregnancy-related mortality
}
\description{
get_ego_age_distribution
}
\section{Details}{

\code{ego_dat} is assumed to have two columns: \code{wwgt} and \code{age.cat}

The age groups used are those returned by
\link{reproductive_age_groups}.
}

