% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dhs_sib_histories.R
\name{get_ego_df}
\alias{get_ego_df}
\title{helper to prep the ego dataset}
\usage{
get_ego_df(df, resp.attrib, verbose = FALSE)
}
\arguments{
\item{df}{the DHS survey dataset}

\item{resp.attrib}{vector with respondent attribute columns (see \link{prep_dhs_sib_histories})}

\item{verbose}{see \link{prep_dhs_sib_histories}}
}
\value{
a prepped ego dataset, used in \link{prep_dhs_sib_histories}
}
\description{
helper to prep the ego dataset
}
