% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dhs_sib_histories.R
\name{prep_dhs_sib_histories}
\alias{prep_dhs_sib_histories}
\title{prepare a DHS dataset for analysis}
\usage{
prep_dhs_sib_histories(df, varmap = sibhist_varmap_dhs6,
  keep_missing = FALSE, keep_varmap_only = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{the raw DHS dataset (indvidual recode)}

\item{varmap}{see Details}

\item{keep_missing}{should we keep reported sibs that are missing sex or survival status?}

\item{keep_varmap_only}{should we only keep ego variables mentioned in the varmap? (Default: FALSE)}

\item{verbose}{report detailed summaries?}
}
\value{
a list; see Details
}
\description{
prepare a DHS dataset for analysis
}
\section{Details}{

\code{varmap} should be a dataframe with columns TODO.
For respondents, you should be sure to include
\itemize{
\item \code{survey} (the survey id, usually a country code plus one digit)
\item \code{caseid} (the respondent id)
\item \code{wwgt} (the sampling weight for women)
\item \code{psu} (the primary sampling unit)
\item \code{doi} (the date of the interview)
}

For siblings, you should be sure to include
\itemize{
\item \code{sib.death.date} (the date of the sibling's death)
\item \code{sib.alive} (whether or not the sib is alive)
\item \code{sib.sex} (the sex of the sibling).
}

The default varmap is \code{sibhist_varmap_dhs6}, which is included with the package.

Returns a list whose entries include
\itemize{
\item \code{ego.dat} - dataset with information about the survey respondents
\item \code{sib.dat} - dataset with information about the reported siblings
\item \code{summ} - a one-row tibble with a summary of the data
}
}

\examples{
  # TODO - write example code
}
