% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling_estimator.R
\name{sibling_estimator}
\alias{sibling_estimator}
\title{Estimate death rates from sibling history data}
\usage{
sibling_estimator(
  sib.dat,
  ego.id,
  sib.id = "sibid",
  sib.frame.indicator,
  sib.sex = "sex",
  cell.config,
  weights,
  boot.weights = NULL,
  return.boot = FALSE,
  discretize.exp = FALSE
)
}
\arguments{
\item{sib.dat}{The long-form sibling history dataset (likely produced by \link{prep_dhs_sib_histories})}

\item{ego.id}{String with the name of the column of \code{sib.dat} that has the ID of the survey respondent}

\item{sib.id}{String with the name of the column of \code{sib.dat} that has the sibling ID. Defaults to \code{'sibid'}, which is the column created by \code{\link{prep_dhs_sib_histories}} and \code{\link{prep_nrsim_sib_histories}}.}

\item{sib.frame.indicator}{String with the name of the column in \code{sib.dat} containing a 0/1 coded variable indicating whether or not each sib is in the frame population}

\item{sib.sex}{String with the name fo the column of \code{sib.dat} that has the sibling's sex}

\item{cell.config}{An object containing the configuration of cells; see TODO for more information}

\item{weights}{String with the name of the column of \code{sib.dat} that has the sampling weight}

\item{boot.weights}{Optional dataframe with bootstrap resampled weights. See Details for more info.}

\item{return.boot}{If TRUE, and if \code{boot.weights} is specified, then return each bootstrap estimate}

\item{discretize.exp}{Boolean for whether or not expsoure should be discretized. Not yet implemented.}
}
\value{
a list with two entries: \code{asdr.ind}, individual visibility asdr estimates; and \code{asdr.agg}, aggregate visibility asdr estimates
}
\description{
Estimate death rates from sibling history data
}
\section{Details}{

If you want estimated sampling variances, you can pass in a data frame \code{boot.weights}.
\code{boot.weights} is assumed to have a column that is named whatever the \code{ego.id} is,
and then a series of columns named \code{boot_weight_1}, ..., \code{boot_weight_M}.
}

