% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling_estimator.R
\name{sibling_estimator}
\alias{sibling_estimator}
\title{Estimate death rates from sibling history data}
\usage{
sibling_estimator(sib.dat, ego.id, sib.id, sib.frame.indicator,
  sib.sex = "sex", cell.config, weights, boot.weights = NULL,
  discretize.exp = FALSE, method = "individual")
}
\arguments{
\item{sib.dat}{The long-form sibling history dataset (likely produced by TODO)}

\item{ego.id}{String with the name of the column of \code{sib.dat} that has the ID of the survey respondent}

\item{sib.id}{String with the name of the column of \code{sib.dat} that has the sibling ID}

\item{sib.frame.indicator}{String with the name of the column in \code{sib.dat} containing a 0/1 coded variable indicating whether or not each sib is in the frame population}

\item{sib.sex}{String with the name fo the column of \code{sib.dat} that has the sibling's sex}

\item{cell.config}{An object containing the configuration of cells; see TODO for more information}

\item{weights}{String with the name of the column of \code{sib.dat} that has the sampling weight}

\item{boot.weights}{Optional dataframe with bootstrap resampled weights. See Details for more info.}

\item{discretize.exp}{Boolean for whether or not expsoure should be discretized. Not yet implemented.}

\item{method}{Which method to use - not yet implemented (currently we just return ind and agg, both without the respondent)}
}
\value{
a list with two entries: \code{asdr.ind}, individual visibility asdr estimates; and \code{asdr.agg}, aggregate visibility asdr estimates
}
\description{
Estimate death rates from sibling history data
}
\section{Details}{

If you want estimated sampling variances, you can pass in a data frame \code{boot.weights}
}

